+++ docs/reference/changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+#### 0.76 (Unreleased)
+
+Misc:
+
+* A new end-to-end test suite setup will help us reduce the cycle time associated with testing Telepresence as we port over existing tests.
+  ([429](https://github.com/datawire/telepresence/pull/429))
+* Improved cleanup of our testing cluster used by CI.
+
 #### 0.75 (January 30, 2018)
 
 Bug fixes: